+++ src/main/java/proteomics/Segment/InferenceSegment.java
@@ -18,24 +18,24 @@ public class InferenceSegment {
     private static final int topNumInEachRegion = 20;
     private static final Pattern pattern = Pattern.compile("([nc][0-9a-i])?([A-Z#$].?)");
 
-    private final float ms2Tolerance;
+    private final double ms2Tolerance;
     private TreeMap<Segment, Integer> aaVectorTemplate = new TreeMap<>();
-    private Map<Float, String> modifiedAAMap = new HashMap<>(35, 1);
-    private final Float[] deltaMassArray;
-    private Map<String, Float> modifiedAAMassMap = new HashMap<>(35, 1);
+    private Map<Double, String> modifiedAAMap = new HashMap<>(35, 1);
+    private final Double[] deltaMassArray;
+    private Map<String, Double> modifiedAAMassMap = new HashMap<>(35, 1);
     private Set<VarModParam> varModParamSet = new HashSet<>();
-    private float[] nTermPossibleMod = null;
-    private float[] cTermPossibleMod = null;
+    private double[] nTermPossibleMod = null;
+    private double[] cTermPossibleMod = null;
     private MassTool massTool;
 
-    public InferenceSegment(MassTool massTool, float ms2Tolerance, Map<String, String> parameterMap, Map<Character, Float> fixModMap) {
+    public InferenceSegment(MassTool massTool, double ms2Tolerance, Map<String, String> parameterMap, Map<Character, Double> fixModMap) {
         this.massTool = massTool;
         this.ms2Tolerance = ms2Tolerance;
-        Map<Character, Float> massTable = massTool.returnMassTable();
+        Map<Character, Double> massTable = massTool.returnMassTable();
 
         char[] standardAaArray = new char[]{'G', 'A', 'S', 'P', 'V', 'T', 'C', 'I', 'L', 'N', 'D', 'Q', 'K', 'E', 'M', 'H', 'F', 'R', 'Y', 'W', 'U', 'O'};
 
-        Map<Float, Character> massAaMap = new HashMap<>(25, 1);
+        Map<Double, Character> massAaMap = new HashMap<>(25, 1);
         for (char aa : standardAaArray) {
             // # = I/L.
             if (aa == 'I' || aa == 'L') {
@@ -66,7 +66,7 @@ public class InferenceSegment {
         }
 
         // generate a mass aa map containing modified amino acid
-        for (float k : massAaMap.keySet()) {
+        for (double k : massAaMap.keySet()) {
             modifiedAAMap.put(k, massAaMap.get(k).toString());
         }
         for (String k : parameterMap.keySet()) {
@@ -74,9 +74,9 @@ public class InferenceSegment {
                 String v = parameterMap.get(k);
                 if (!v.startsWith("0.0")) {
                     String[] temp = v.split("@");
-                    float tempMass = massTable.get(temp[1].charAt(0)) + Float.valueOf(temp[0]);
+                    double tempMass = massTable.get(temp[1].charAt(0)) + Double.valueOf(temp[0]);
                     // check if the mass has conflict
-                    for (float temp2 : modifiedAAMap.keySet()) {
+                    for (double temp2 : modifiedAAMap.keySet()) {
                         if (Math.abs(temp2 - tempMass) <= ms2Tolerance) {
                             logger.error("{} and {} have conflict mass values({} vs {}).", v, modifiedAAMap.get(temp2), tempMass, temp2);
                             System.exit(1);
@@ -86,12 +86,12 @@ public class InferenceSegment {
                         // fix modification and var modification cannot be coexist
                         if ((temp[1].charAt(0) == 'I') || (temp[1].charAt(0) == 'L')) {
                             modifiedAAMap.put(tempMass, temp[1].replace(temp[1].charAt(0), '#'));
-                            modifiedAAMassMap.put(temp[1].replace(temp[1].charAt(0), '#'), Float.valueOf(temp[0]));
+                            modifiedAAMassMap.put(temp[1].replace(temp[1].charAt(0), '#'), Double.valueOf(temp[0]));
                         } else {
                             modifiedAAMap.put(tempMass, temp[1]);
-                            modifiedAAMassMap.put(temp[1], Float.valueOf(temp[0]));
+                            modifiedAAMassMap.put(temp[1], Double.valueOf(temp[0]));
                         }
-                        varModParamSet.add(new VarModParam(Float.valueOf(temp[0]), temp[1].charAt(0), 1, false)); // var mods from the parameter file have the highest priority, those PTM can exist in peptide terminal.
+                        varModParamSet.add(new VarModParam(Double.valueOf(temp[0]), temp[1].charAt(0), 1, false)); // var mods from the parameter file have the highest priority, those PTM can exist in peptide terminal.
                     }
                 }
             } else if (k.contentEquals("Nterm")) {
@@ -99,10 +99,10 @@ public class InferenceSegment {
                     // fix modification and var modification cannot be coexist
                     if (!parameterMap.get(k).startsWith("0.0")) {
                         String[] tempArray = parameterMap.get(k).split(",");
-                        nTermPossibleMod = new float[tempArray.length];
+                        nTermPossibleMod = new double[tempArray.length];
                         for (int i = 0; i < tempArray.length; ++i) {
-                            nTermPossibleMod[i] = Float.valueOf(tempArray[i].trim());
-                            varModParamSet.add(new VarModParam(Float.valueOf(tempArray[i].trim()), 'n', 1, false)); // var mods from the parameter file have the highest priority
+                            nTermPossibleMod[i] = Double.valueOf(tempArray[i].trim());
+                            varModParamSet.add(new VarModParam(Double.valueOf(tempArray[i].trim()), 'n', 1, false)); // var mods from the parameter file have the highest priority
                         }
                     }
                 }
@@ -111,19 +111,19 @@ public class InferenceSegment {
                     // fix modification and var modification cannot be coexist
                     if (!parameterMap.get(k).startsWith("0.0")) {
                         String[] tempArray = parameterMap.get(k).split(",");
-                        cTermPossibleMod = new float[tempArray.length];
+                        cTermPossibleMod = new double[tempArray.length];
                         for (int i = 0; i < tempArray.length; ++i) {
-                            cTermPossibleMod[i] = Float.valueOf(tempArray[i].trim());
-                            varModParamSet.add(new VarModParam(Float.valueOf(tempArray[i].trim()), 'c', 1, false)); // var mods from the parameter file have the highest priority
+                            cTermPossibleMod[i] = Double.valueOf(tempArray[i].trim());
+                            varModParamSet.add(new VarModParam(Double.valueOf(tempArray[i].trim()), 'c', 1, false)); // var mods from the parameter file have the highest priority
                         }
                     }
                 }
             }
         }
-        deltaMassArray = modifiedAAMap.keySet().toArray(new Float[modifiedAAMap.size()]);
+        deltaMassArray = modifiedAAMap.keySet().toArray(new Double[modifiedAAMap.size()]);
     }
 
-    public List<ThreeExpAA> inferSegmentLocationFromSpectrum(float precursorMass, TreeMap<Float, Float> plMap) {
+    public List<ThreeExpAA> inferSegmentLocationFromSpectrum(double precursorMass, TreeMap<Double, Double> plMap) {
         return inferThreeAAFromSpectrum(addVirtualPeaks(precursorMass, plMap), precursorMass - massTool.H2O + MassTool.PROTON);
     }
 
@@ -133,9 +133,9 @@ public class InferenceSegment {
             return finalVector;
         } else {
             for (ThreeExpAA expAaList : inputList) {
-                float totalIntensity = expAaList.getTotalIntensity();
+                double totalIntensity = expAaList.getTotalIntensity();
                 int idx = aaVectorTemplate.get(new Segment(expAaList.getPtmFreeAAString()));
-                float value = Math.max(totalIntensity, finalVector.get(idx));
+                double value = Math.max(totalIntensity, finalVector.get(idx));
                 finalVector.put(idx, value);
             }
             return finalVector;
@@ -151,7 +151,7 @@ public class InferenceSegment {
         return new SparseBooleanVector(tempSet);
     }
 
-    public Map<String, Float> getModifiedAAMassMap() {
+    public Map<String, Double> getModifiedAAMassMap() {
         return modifiedAAMassMap;
     }
 
@@ -159,11 +159,11 @@ public class InferenceSegment {
         return varModParamSet;
     }
 
-    public float[] getNTermPossibleMod() {
+    public double[] getNTermPossibleMod() {
         return nTermPossibleMod;
     }
 
-    public float[] getCTermPossibleMod() {
+    public double[] getCTermPossibleMod() {
         return cTermPossibleMod;
     }
 
@@ -171,23 +171,23 @@ public class InferenceSegment {
         return seq.replaceAll("[IL]", "#");
     }
 
-    private List<ThreeExpAA> inferThreeAAFromSpectrum(TreeMap<Float, Float> plMap, float cTermMz) {
-        Float[] mzArray = plMap.keySet().toArray(new Float[plMap.size()]);
-        Float[] intensityArray = plMap.values().toArray(new Float[plMap.size()]);
+    private List<ThreeExpAA> inferThreeAAFromSpectrum(TreeMap<Double, Double> plMap, double cTermMz) {
+        Double[] mzArray = plMap.keySet().toArray(new Double[plMap.size()]);
+        Double[] intensityArray = plMap.values().toArray(new Double[plMap.size()]);
         Set<ThreeExpAA> tempSet = new HashSet<>();
         List<ThreeExpAA> outputList = new LinkedList<>();
         for (int i = 0; i < mzArray.length; ++i) {
-            float mz1 = mzArray[i];
-            float intensity1 = intensityArray[i];
+            double mz1 = mzArray[i];
+            double intensity1 = intensityArray[i];
             for (int j = i + 1; j < mzArray.length; ++j) {
-                float mz2 = mzArray[j];
-                float intensity2 = intensityArray[j];
+                double mz2 = mzArray[j];
+                double intensity2 = intensityArray[j];
                 String aa1 = inferAA(mz1, mz2, Math.abs(mz1 - MassTool.PROTON) <= ms2Tolerance, false);
                 if (aa1 != null) {
                     Matcher matcher = pattern.matcher(aa1);
                     char ptmFreeAA = '\0';
-                    float mod = 0;
-                    float nTermMod = 0;
+                    double mod = 0;
+                    double nTermMod = 0;
                     if (matcher.matches()) {
                         if (modifiedAAMassMap.containsKey(matcher.group(2))) {
                             mod = modifiedAAMassMap.get(matcher.group(2));
@@ -208,8 +208,8 @@ public class InferenceSegment {
                     ExpAA expAa1 = new ExpAA(aa1, ptmFreeAA, mz1, mz2, intensity1, intensity2, -1, mod, nTermMod, 0);
                     List<List<ExpAA>> tempAasList2 = new LinkedList<>();
                     for (int k = j + 1; k < mzArray.length; ++k) {
-                        float mz3 = mzArray[k];
-                        float intensity3 = intensityArray[k];
+                        double mz3 = mzArray[k];
+                        double intensity3 = intensityArray[k];
                         String aa2 = inferAA(mz2, mz3, false, false);
                         if (aa2 != null) {
                             mod = 0;
@@ -219,14 +219,14 @@ public class InferenceSegment {
                             ExpAA expAa2 = new ExpAA(aa2, aa2.charAt(0), mz2, mz3, intensity2, intensity3, -1, mod, 0, 0);
                             List<ExpAA> tempAasList3 = new LinkedList<>();
                             for (int l = k + 1; l < mzArray.length; ++l) {
-                                float mz4 = mzArray[l];
-                                float intensity4 = intensityArray[l];
+                                double mz4 = mzArray[l];
+                                double intensity4 = intensityArray[l];
                                 String aa3 = inferAA(mz3, mz4, false, Math.abs(mz4 - cTermMz) <= ms2Tolerance);
                                 if (aa3 != null) {
                                     matcher = pattern.matcher(aa3);
                                     ptmFreeAA = '\0';
                                     mod = 0;
-                                    float cTermMod = 0;
+                                    double cTermMod = 0;
                                     if (matcher.matches()) {
                                         if (modifiedAAMassMap.containsKey(matcher.group(2))) {
                                             mod = modifiedAAMassMap.get(matcher.group(2));
@@ -285,21 +285,21 @@ public class InferenceSegment {
         }
 
         if (tempList.size() > minTagNum) {
-            float minMz = plMap.firstKey();
-            float regionWindow = (float) Math.ceil((plMap.lastKey() - minMz) / regionNum);
+            double minMz = plMap.firstKey();
+            double regionWindow = Math.ceil((plMap.lastKey() - minMz) / regionNum);
             for (ThreeExpAA expAa : tempList) {
                 expAa.setRegionIdx((int) Math.floor((expAa.getHeadLocation() - minMz) / regionWindow));
             }
-            List<List<Float>> regionIntensityList = new ArrayList<>(20);
+            List<List<Double>> regionIntensityList = new ArrayList<>(20);
             for (int i = 0; i < regionNum; ++i) {
                 regionIntensityList.add(new ArrayList<>(100));
             }
             for (ThreeExpAA expAa : tempList) {
                 regionIntensityList.get(expAa.getRegionIdx()).add(expAa.getTotalIntensity());
             }
-            float[] intensityTArray = new float[regionNum];
+            double[] intensityTArray = new double[regionNum];
             for (int i = 0; i < regionNum; ++i) {
-                List<Float> intensityList = regionIntensityList.get(i);
+                List<Double> intensityList = regionIntensityList.get(i);
                 Collections.sort(intensityList, Collections.reverseOrder());
                 if (intensityList.size() > topNumInEachRegion) {
                     intensityTArray[i] = intensityList.get(topNumInEachRegion);
@@ -316,16 +316,16 @@ public class InferenceSegment {
         }
     }
 
-    private String inferAA(float mz1, float mz2, boolean nTerm, boolean cTerm) {
-        float mzDiff = mz2 - mz1;
-        for (float mass : deltaMassArray) {
+    private String inferAA(double mz1, double mz2, boolean nTerm, boolean cTerm) {
+        double mzDiff = mz2 - mz1;
+        for (double mass : deltaMassArray) {
             if (Math.abs(mzDiff - mass) <= 2 * ms2Tolerance) {
                 return modifiedAAMap.get(mass);
             }
         }
 
         if (nTerm && (nTermPossibleMod != null)) {
-            for (float mass : deltaMassArray) {
+            for (double mass : deltaMassArray) {
                 for (int i = 0; i < nTermPossibleMod.length; ++i) {
                     if (Math.abs(mzDiff - mass - nTermPossibleMod[i]) <= 2 * ms2Tolerance) {
                         return "n" + i + modifiedAAMap.get(mass);
@@ -335,7 +335,7 @@ public class InferenceSegment {
         }
 
         if (cTerm && (cTermPossibleMod != null)) {
-            for (float mass : deltaMassArray) {
+            for (double mass : deltaMassArray) {
                 for (int i = 0; i < cTermPossibleMod.length; ++i) {
                     if (Math.abs(mzDiff - mass - cTermPossibleMod[i]) <= 2 * ms2Tolerance) {
                         return "c" + i + modifiedAAMap.get(mass);
@@ -347,17 +347,17 @@ public class InferenceSegment {
         return null;
     }
 
-    private TreeMap<Float, Float> addVirtualPeaks(float precursorMass, TreeMap<Float, Float> plMap) {
-        float totalMass = precursorMass + 2 * MassTool.PROTON;
-        TreeMap<Float, Float> finalPlMap = new TreeMap<>();
-        for (float mz : plMap.keySet()) {
+    private TreeMap<Double, Double> addVirtualPeaks(double precursorMass, TreeMap<Double, Double> plMap) {
+        double totalMass = precursorMass + 2 * MassTool.PROTON;
+        TreeMap<Double, Double> finalPlMap = new TreeMap<>();
+        for (double mz : plMap.keySet()) {
             finalPlMap.put(mz, plMap.get(mz));
         }
-        for (float mz : plMap.keySet()) {
-            float anotherMz = totalMass - mz;
-            float leftMz = anotherMz - ms2Tolerance;
-            float rightMz = anotherMz + ms2Tolerance;
-            NavigableMap<Float, Float> temp = null;
+        for (double mz : plMap.keySet()) {
+            double anotherMz = totalMass - mz;
+            double leftMz = anotherMz - ms2Tolerance;
+            double rightMz = anotherMz + ms2Tolerance;
+            NavigableMap<Double, Double> temp = null;
             try {
                 temp = plMap.subMap(leftMz, true, rightMz, true);
             } catch (IllegalArgumentException ex) {}
@@ -368,16 +368,16 @@ public class InferenceSegment {
         }
 
         // Add two virtual peak. Because we have convert all y-ions to b-ions.
-        finalPlMap.put(MassTool.PROTON, 1f);
-        float cTermMz = precursorMass - massTool.H2O + MassTool.PROTON;
-        float leftMz = cTermMz - ms2Tolerance;
-        float rightMz = cTermMz + ms2Tolerance;
-        NavigableMap<Float, Float> temp = null;
+        finalPlMap.put(MassTool.PROTON, 1d);
+        double cTermMz = precursorMass - massTool.H2O + MassTool.PROTON;
+        double leftMz = cTermMz - ms2Tolerance;
+        double rightMz = cTermMz + ms2Tolerance;
+        NavigableMap<Double, Double> temp = null;
         try {
             temp = plMap.subMap(leftMz, true, rightMz, true);
         } catch (IllegalArgumentException ex) {}
         if ((temp == null) || (temp.isEmpty())) {
-            finalPlMap.put(cTermMz, 1f);
+            finalPlMap.put(cTermMz, 1d);
         }
 
         return finalPlMap;
